Projectile Impact Niagra Particle System issue
Prototyping the different particle systems of Niagara I'm running into the issue of the Particle not initiating on impact of the projectile on the ForceField. The tutorials work from verions 4 but there are issues online with version 5 and the particle systems not working in blurprints.

But so far, the bounce effect is working. I've also duplicated the mesh and applied a M_Floor texture to see the duplicated mesh. Then I cut the top of the duplicated ForceField to create a impact outline on the mesh. In the material graph I added a particle colour, radial gradient exponential and time to create an effect of lines overlapping the cut mesh.