Impact VFX Working
I found a post about applying the Nigara particle system to the System Template in the blueprint of the object. This fixed the issue, I just overlooked where to enable the particle system. Now, on projectile location and rotation the hit projectile makes a 'fountain' burst effect, bounced effect and overlayed impact effects.