Finishing touches
I added a ground Nigara particle system to finish the look of the shield. ( To match the nonclipping effect of the outside shield, I added a SceneTexture:WorldNormal for the system to recognise the collision of the walls and activate a Mask (absolute). Before, the particle would pull upwards against the wall so this fixes that.