Fracturing Geometry Collections experiment
Looking at different VFX outcomes of projectile impact on the forcefield, I want to attempt to have a health system applied to the shield so it will fracture or break when health is 0. Before that, I tested out the Fracture feature of unreal. Selecting a new shpere and applying the ForceField Inst to it. Then, I apply a Uniform Fracture (Min and Max set to 80) and changed the mass to 100. ( So the destruction physics is aligned with the power of the Projectile) For the projectile to break the object, I needed to scale up the BP_ShooterProjectile sphere up from 0.5 to 1.5. ( The physics of the fracture is only applied to a certain amount of force )