Texture Update using Shield Health WORKING
Applied a CurrentHealth to three textures that change depending on the damage. -1 on impact, texture 1-3 would simulate. Once the health reaches 0, the object would be destroyed. I want to apply the fracture and break with the Chaos destruction, therefore the fractures would be simulated without textures, and the forcefield Niagara destruction would play once the forcefield's health reaches 0.