Decals with render target Prototype
With the Glass health applied to a new Box object, I want to add Decal textures to the M_Glass material and activate it with the BP_Glass blueprint so that the masked texture (Projectile impact/Glass shatter) will 'delete' the space made by the fracture/impact. I will have to download resources to see how the Decal is applied.