Prototype 3 Fracture system with VFX Impact + Fixing up Comments on code
Assigned the Geometry Collection "Rest collection" mesh to the M_Glass ( Where it has the fracture + Bounce VFX ) and the conjunction of the Chaos destruction worked with the VFX effects of the force field. Beyond this, I will add another prototype to experiment with the Decal textures + Assigning it to the world position of the hit to match the impact to the fracture.